moved document scoring to separate module
